+++ category/src/main/java/com/moa/CategoryApplication.java
@@ -1,0 +1,13 @@
+package com.moa;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CategoryApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CategoryApplication.class, args);
+    }
+
+}
