--- category/src/main/java/com/moa/category/CategoryApplication.java
@@ -1,13 +1,0 @@
-package com.moa.category;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CategoryApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CategoryApplication.class, args);
-	}
-
-}
