+++ category/src/main/java/com/moa/category/application/CategoryService.java
@@ -1,0 +1,4 @@
+package com.moa.category.application;
+
+public interface CategoryService {
+}
