+++ category/src/main/java/com/moa/category/application/CategoryServiceImpl.java
@@ -1,0 +1,4 @@
+package com.moa.category.application;
+
+public class CategoryServiceImpl {
+}
