--- category/src/main/java/com/moa/category/category/application/CategoryServiceImpl.java
@@ -1,4 +1,0 @@
-package com.moa.category.category.application;
-
-public class CategoryServiceImpl {
-}
