--- category/src/main/java/com/moa/category/category/infrastructure/UserInterestRepository.java
@@ -1,7 +1,0 @@
-package com.moa.category.category.infrastructure;
-
-import com.moa.category.category.domain.UserInterest;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserInterestRepository extends JpaRepository<UserInterest,Long> {
-}
