--- category/src/main/java/com/moa/category/global/domain/BaseEnum.java
@@ -1,9 +1,0 @@
-package com.moa.category.global.domain;
-
-
-public interface BaseEnum<T, K> {
-
-	T getCode();
-	K getTitle();
-
-}
