--- category/src/main/java/com/moa/category/global/domain/CustomException.java
@@ -1,14 +1,0 @@
-package com.moa.category.global.domain;
-
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-
-@Getter
-@RequiredArgsConstructor
-public class CustomException extends RuntimeException {
-
-	private final ErrorCode errorCode;
-
-}
