+++ category/src/main/java/com/moa/category/infrastructure/CategoryMeetingListDto.java
@@ -1,0 +1,8 @@
+package com.moa.category.infrastructure;
+
+import com.moa.category.domain.CategoryMeetingList;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CategoryMeetingListDto extends JpaRepository<CategoryMeetingList,Long>{
+
+}
