+++ category/src/main/java/com/moa/category/infrastructure/ThemeCategoryRepository.java
@@ -1,0 +1,7 @@
+package com.moa.category.infrastructure;
+
+import com.moa.category.domain.ThemeCategory;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ThemeCategoryRepository extends JpaRepository<ThemeCategory,Integer> {
+}
