+++ category/src/main/java/com/moa/category/infrastructure/UserInterestRepository.java
@@ -1,0 +1,7 @@
+package com.moa.category.infrastructure;
+
+import com.moa.category.domain.UserInterest;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserInterestRepository extends JpaRepository<UserInterest,Long> {
+}
