+++ category/src/main/java/com/moa/global/domain/BaseEnum.java
@@ -1,0 +1,9 @@
+package com.moa.category.global.domain;
+
+
+public interface BaseEnum<T, K> {
+
+    T getCode();
+    K getTitle();
+
+}
