+++ category/src/main/java/com/moa/global/domain/CustomException.java
@@ -1,0 +1,14 @@
+package com.moa.global.domain;
+
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+
+@Getter
+@RequiredArgsConstructor
+public class CustomException extends RuntimeException {
+
+    private final ErrorCode errorCode;
+
+}
