--- category/src/test/java/com/moa/category/CategoryApplicationTests.java
@@ -1,13 +1,0 @@
-package com.moa.category;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CategoryApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
